Separated import and command files for Equinix
